# Improve blob unit test performance